server: Move unsafeHashByRev to new hash.go file
